@@ -2,118 +2,84 @@ package org.imaginativeworld.oopsnointernet.sample
 
 
 import android.content.Intent
+import android.content.res.Configuration
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import kotlinx.android.synthetic.main.activity_main.*
-import org.imaginativeworld.oopsnointernet.ConnectionCallback
-import org.imaginativeworld.oopsnointernet.NoInternetDialog
-import org.imaginativeworld.oopsnointernet.NoInternetSnackbar
+import androidx.appcompat.app.AppCompatDelegate
+import androidx.core.content.ContextCompat
+import org.imaginativeworld.oopsnointernet.sample.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
 
-    // No Internet Dialog
-    private var noInternetDialog: NoInternetDialog? = null
+    private lateinit var binding: ActivityMainBinding
 
-    // No Internet Snackbar
-    private var noInternetSnackbar: NoInternetSnackbar? = null
-
-    private var selectedRadioId = R.id.radio_dialog
+    private var selectedType = Constants.TYPE_DIALOG_PENDULUM
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
-        radio_group.setOnCheckedChangeListener { _, checkedId ->
-
-            selectedRadioId = checkedId
-
-            init()
-
+        // Init views
+        if (isNightMode()) {
+            binding.fabThemeToggle.setImageDrawable(
+                ContextCompat.getDrawable(
+                    this,
+                    R.drawable.ic_round_bedtime_24
+                )
+            )
+        } else {
+            binding.fabThemeToggle.setImageDrawable(
+                ContextCompat.getDrawable(
+                    this,
+                    R.drawable.ic_round_wb_sunny_24
+                )
+            )
         }
 
-        btn_go_to_java_activity.setOnClickListener {
+        // Init listeners
+        binding.radioGroup.setOnCheckedChangeListener { _, checkedId ->
 
-            startActivity(Intent(this, JavaActivity::class.java))
-
-        }
-    }
-
-    override fun onResume() {
-        super.onResume()
-
-        init()
-    }
-
-    private fun init() {
-        when (selectedRadioId) {
-
-            R.id.radio_dialog -> {
-
-                noInternetSnackbar?.destroy()
-
-                // No Internet Dialog
-                noInternetDialog = NoInternetDialog.Builder(this)
-                    .apply {
-                        connectionCallback = object : ConnectionCallback { // Optional
-                            override fun hasActiveConnection(hasActiveConnection: Boolean) {
-                                // ...
-                            }
-                        }
-                        cancelable = false // Optional
-                        noInternetConnectionTitle = "No Internet" // Optional
-                        noInternetConnectionMessage =
-                            "Check your Internet connection and try again." // Optional
-                        showInternetOnButtons = true // Optional
-                        pleaseTurnOnText = "Please turn on" // Optional
-                        wifiOnButtonText = "Wifi" // Optional
-                        mobileDataOnButtonText = "Mobile data" // Optional
-
-                        onAirplaneModeTitle = "No Internet" // Optional
-                        onAirplaneModeMessage = "You have turned on the airplane mode." // Optional
-                        pleaseTurnOffText = "Please turn off" // Optional
-                        airplaneModeOffButtonText = "Airplane mode" // Optional
-                        showAirplaneModeOffButtons = true // Optional
-                    }
-                    .build()
-
-            }
-
-            R.id.radio_snackbar -> {
-
-                noInternetDialog?.destroy()
-
-                // No Internet Snackbar
-                noInternetSnackbar =
-                    NoInternetSnackbar.Builder(this, findViewById(android.R.id.content))
-                        .apply {
-                            connectionCallback = object : ConnectionCallback { // Optional
-                                override fun hasActiveConnection(hasActiveConnection: Boolean) {
-                                    // ...
-                                }
-                            }
-                            indefinite = true // Optional
-                            noInternetConnectionMessage =
-                                "No active Internet connection!" // Optional
-                            onAirplaneModeMessage =
-                                "You have turned on the airplane mode!" // Optional
-                            snackbarActionText = "Settings" // Optional
-                            showActionToDismiss = false // Optional
-                            snackbarDismissActionText = "OK" // Optional
-                        }
-                        .build()
-
+            if (checkedId == R.id.radio_dialog_pendulum) {
+                selectedType = Constants.TYPE_DIALOG_PENDULUM
+            } else if (checkedId == R.id.radio_dialog_signal) {
+                selectedType = Constants.TYPE_DIALOG_SIGNAL
+            } else if (checkedId == R.id.radio_snackbar_fire) {
+                selectedType = Constants.TYPE_SNACKBAR_FIRE
             }
 
         }
+
+        binding.fabThemeToggle.setOnClickListener {
+
+            if (isNightMode()) {
+                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
+            } else {
+                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
+            }
+
+        }
+
+        binding.btnKotlin.setOnClickListener {
+
+            startActivity(Intent(this, KotlinExampleActivity::class.java)
+                .apply {
+                    this.putExtra(Constants.KEY_TYPE, selectedType)
+                })
+
+        }
+
+        binding.btnJava.setOnClickListener {
+
+            startActivity(Intent(this, JavaExampleActivity::class.java)
+                .apply {
+                    this.putExtra(Constants.KEY_TYPE, selectedType)
+                })
+
+        }
     }
 
-    override fun onPause() {
-        super.onPause()
-
-        // No Internet Dialog
-        noInternetDialog?.destroy()
-
-        // No Internet Snackbar
-        noInternetSnackbar?.destroy()
+    private fun isNightMode(): Boolean {
+        return resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK == Configuration.UI_MODE_NIGHT_YES
     }
 }
